requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.29.0
attrs==24.2.0
certifi==2023.11.17
et_xmlfile==2.0.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
psycopg==3.2.5
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.11.1
pytz==2024.2
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2024.2
wheel==0.45.1
xlrd==2.0.1
yarl==1.13.1
